morph-kgc compatibility
The functionality of extracting triples from XML files in RMLMappings has been replaced by using the morph-kgc library, which is faster in this task. For this purpose an xml2rdf_morph script has been included in addition to the original xml2rdf. The morph functionality is tested in the python notebook rmlmappingToMorph.ipynb. It has been necessary to correct the .ttl mapping files opencorporates_mapping.ttl and openopps_mapping.ttl, keeping the old ones in the directory knowledge-graphrml-mappings_mappings_old_mappings.